+++ src/client/features/available-golfers/components/checkbox.tsx
@@ -1,6 +1,6 @@
 import { Checkbox as MUICheckbox } from '@mui/material';
 import React from 'react';
-import { StyledFormControlLabel } from '../styled';
+import { StyledFormControlLabel } from './styled';
 
 interface Props {
   label: string;
