+++ src/client/features/available-golfers/components/search-bar.tsx
@@ -2,7 +2,7 @@ import CloseIcon from '@mui/icons-material/Close';
 import SearchIcon from '@mui/icons-material/Search';
 import { InputAdornment } from '@mui/material';
 import React from 'react';
-import { SearchBarWrapper, StyledIconButton, StyledSearchBar } from '../styled';
+import { SearchBarWrapper, StyledIconButton, StyledSearchBar } from './styled';
 
 interface Props {
   searchTerm: string;
