+++ src/client/features/team/components/team-page.tsx
@@ -3,10 +3,10 @@ import { useSelector } from 'react-redux';
 import { useGetTeam } from 'src/client/api';
 import { Loader } from 'src/client/features/shared';
 import { selectPhaseSelection } from 'src/client/store';
+import { AvailableGolfersList } from '../../available-golfers/components/available-golfers';
 import { Login } from '../../login/components/login';
 import { useGoogleSignIn } from '../../login/hooks/use-google-sign-in';
-import { AvailableGolfersList } from './available-golfers/available-golfers';
-import { SelectedTeam } from './selected-team/selected-team';
+import { SelectedTeam } from '../../selected-team/components/selected-team';
 import { TeamPageContainer } from './styled';
 
 export const TeamPage = () => {
